Fixed build issue

+++ app/Http/Controllers/Api/AdminSellerController.php
@@ -6,10 +6,13 @@ use App\Http\Controllers\Controller;
 use App\Services\Admin\SellerService;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rules\Password;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class AdminSellerController extends Controller
 {
     protected \App\Services\Admin\SellerService $service;
+    const MESSAGE = '403 Forbidden';
 
     public function __construct(SellerService $service)
     {
@@ -18,25 +21,28 @@ class AdminSellerController extends Controller
 
     public function allSellers(): array
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         return $this->service->allSellers();
     }
 
     public function approveSeller(Request $request)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         $request->validate([
             'user_id' => ['required', 'integer', 'exists:users,id']
         ]);
-
         return $this->service->approveSeller($request);
     }
 
     public function viewSeller($id)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         return $this->service->viewSeller($id);
     }
 
     public function editSeller(Request $request, $id)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         $request->validate([
             'first_name' => ['string', 'max:255'],
             'last_name' => ['string', 'max:255'],
@@ -50,6 +56,7 @@ class AdminSellerController extends Controller
 
     public function banSeller(Request $request)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         $request->validate([
             'user_id' => ['required', 'integer', 'exists:users,id']
         ]);
@@ -59,16 +66,19 @@ class AdminSellerController extends Controller
 
     public function removeSeller($id)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         return $this->service->removeSeller($id);
     }
 
     public function paymentHistory($id)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         return $this->service->paymentHistory($id);
     }
 
     public function bulkRemove(Request $request)
     {
+        abort_if(Gate::denies('seller_management'), Response::HTTP_FORBIDDEN, self::MESSAGE);
         $request->validate([
             'user_ids' => 'required|array',
             'user_ids.*' => 'exists:users,id',
